+++ bespokeasm/__main__.py
@@ -19,6 +19,11 @@ def main():
 @click.argument('asm_file')
 @click.option('--config-file', '-c', required=True, help='The filepath to the instruction set configuration file,')
 @click.option(
+        '--generate-binary/--no-binary',
+        default=True,
+        help='Indicates whether a binary image of the compiled bytecode should be generated.'
+    )
+@click.option(
         '--output-file', '-o',
         help='The filepath to where the binary image will be written. Defaults to '
              'the input file name with a *.bin extension.'
@@ -39,9 +44,16 @@ def main():
              'a specific size.'
     )
 @click.option(
-        '--pretty-print', '-p', is_flag=True, default=False,
+        '--pretty-print', '-p',
+        is_flag=True, default=False,
         help='if true, a pretty print version of the compilation will be produced.'
     )
+@click.option(
+        '--pretty-print-format', '-t',
+        type=click.Choice(['source_details', 'minhex'], case_sensitive=False),
+        default='source_details',
+        help='The format that should be used when pretty printing.',
+)
 @click.option(
         '--pretty-print-output',  default='stdout',
         help='if pretty-print is enabled, this specifies the output file. Defaults to stdout.'
@@ -54,11 +66,13 @@ def main():
 def compile(
             asm_file,
             config_file,
+            generate_binary,
             output_file,
             binary_min_address,
             binary_max_address,
             binary_fill,
             pretty_print,
+            pretty_print_format,
             pretty_print_output,
             verbose,
             include_path
@@ -67,17 +81,19 @@ def compile(
         output_file = os.path.splitext(asm_file)[0] + '.bin'
     if verbose:
         click.echo(f'The file to assemble is: {asm_file}')
-        click.echo(f'The binary image will be written to: {output_file}')
+        if generate_binary:
+            click.echo(f'The binary image will be written to: {output_file}')
         if int(binary_min_address) > 0:
             click.echo(f'  with the starting address written: {binary_min_address}')
         if int(binary_max_address) >= 0:
             click.echo(f'  with the maximum address written: {binary_max_address}')
 
     asm = Assembler(
-        asm_file, config_file, output_file,
+        asm_file, config_file,
+        generate_binary, output_file,
         int(binary_min_address), int(binary_max_address) if int(binary_max_address) >= 0 else None,
         binary_fill,
-        pretty_print, pretty_print_output, verbose,
+        pretty_print, pretty_print_format, pretty_print_output, verbose,
         include_path,
     )
     asm.assemble_bytecode()