+++ bespokeasm/assembler/engine.py
@@ -12,18 +12,21 @@ from bespokeasm.assembler.line_object.label_line import LabelLine
 from bespokeasm.assembler.line_object.predefined_data import PredefinedDataLine
 from bespokeasm.assembler.model import AssemblerModel
 from bespokeasm.assembler.label_scope import LabelScope
-
+from bespokeasm.assembler.pretty_printer import PrettyPrinterBase
+from bespokeasm.assembler.pretty_printer.factory import PrettyPrinterFactory
 
 class Assembler:
     def __init__(
                 self,
                 source_file: str,
                 config_file: str,
+                generate_binary: bool,
                 output_file: str,
                 binary_start: int,
                 binary_end: int,
                 binary_fill_value: int,
                 enable_pretty_print: bool,
+                pretty_print_format: str,
                 pretty_print_output: str,
                 is_verbose: int,
                 include_paths: list[str],
@@ -31,7 +34,9 @@ class Assembler:
         self.source_file = source_file
         self._output_file = output_file
         self._config_file = config_file
+        self._generate_binary = generate_binary
         self._enable_pretty_print = enable_pretty_print
+        self._pretty_print_format = pretty_print_format
         self._pretty_print_output = pretty_print_output
         self._binary_fill_value = binary_fill_value & 0xff
         self._verbose = is_verbose
@@ -110,88 +115,35 @@ class Assembler:
         fill_bytes = bytearray([self._binary_fill_value])
         addr = self._binary_start
 
-        if self._verbose > 2:
-            print("\nGenerating byte code:")
-        while addr <= (max_generated_address if self._binary_end is None else self._binary_end):
-            lobj = line_dict.get(addr, None)
-            insertion_bytes = fill_bytes
-            if lobj is not None:
-                line_bytes = lobj.get_bytes()
-                if line_bytes is not None:
-                    insertion_bytes = line_bytes
-                    if self._verbose > 2:
-                        line_bytes_str = binascii.hexlify(line_bytes, sep=' ').decode("utf-8")
-                        click.echo(f'Address ${addr:x} : {lobj} bytes = {line_bytes_str}')
-            byte_code.extend(insertion_bytes)
-            addr += len(insertion_bytes)
+        if self._generate_binary:
+            if self._verbose > 2:
+                print("\nGenerating byte code:")
+            while addr <= (max_generated_address if self._binary_end is None else self._binary_end):
+                lobj = line_dict.get(addr, None)
+                insertion_bytes = fill_bytes
+                if lobj is not None:
+                    line_bytes = lobj.get_bytes()
+                    if line_bytes is not None:
+                        insertion_bytes = line_bytes
+                        if self._verbose > 2:
+                            line_bytes_str = binascii.hexlify(line_bytes, sep=' ').decode("utf-8")
+                            click.echo(f'Address ${addr:x} : {lobj} bytes = {line_bytes_str}')
+                byte_code.extend(insertion_bytes)
+                addr += len(insertion_bytes)
 
-        click.echo(f'Writing {len(byte_code)} bytes of byte code to {self._output_file}')
-        with open(self._output_file, 'wb') as f:
-            f.write(byte_code)
+            click.echo(f'Writing {len(byte_code)} bytes of byte code to {self._output_file}')
+            with open(self._output_file, 'wb') as f:
+                f.write(byte_code)
+        elif self._verbose > 1:
+            print('NOT writing byte code to binary image.')
 
         if self._enable_pretty_print:
-            pretty_str = self._pretty_print_results(line_obs, max_instruction_text_size)
+            pprinter = PrettyPrinterFactory.getPrettyPrinter(self._pretty_print_format, line_obs, self._model)
+            pretty_str = pprinter.pretty_print(max_instruction_text_size)
             if self._pretty_print_output == 'stdout':
                 print(pretty_str)
             else:
                 with open(self._pretty_print_output, 'w') as f:
                     f.write(pretty_str)
 
-    def _pretty_print_results(self, line_obs, max_instruction_text_size):
-        output = io.StringIO()
 
-        address_size = math.ceil(self._model.address_size/4)
-        address_format_str = f'0x{{0:0{address_size}x}}'
-        COL_WIDTH_LINE = 7
-        COL_WIDTH_ADDRESS = max(address_size + 3, 7)
-        COL_WIDTH_BYTE = 4
-        COL_WIDTH_BINARY = 8
-        blank_line_num = ''.join([' '*COL_WIDTH_LINE])
-        blank_instruction_text = ''.join([' '*max_instruction_text_size])
-        blank_address_text = ''.join([' '*COL_WIDTH_ADDRESS])
-        blank_byte_text = ''.join([' '*COL_WIDTH_BYTE])
-        blank_binary_text = ''.join([' '*COL_WIDTH_BINARY])
-
-        header_text = ' {0} | {1} | {2} | {3} | {4} | Comment '.format(
-            'Line'.center(COL_WIDTH_LINE),
-            'Code'.ljust(max_instruction_text_size),
-            'Address'.center(COL_WIDTH_ADDRESS),
-            'Byte'.center(COL_WIDTH_BYTE),
-            'Binary'.center(COL_WIDTH_BINARY),
-        )
-        header_line_text = '-{0}-+-{1}-+-{2}-+-{3}-+-{4}-+---------------'.format(
-            ''.join('-'*(COL_WIDTH_LINE)),
-            ''.join('-'*(max_instruction_text_size)),
-            ''.join('-'*(COL_WIDTH_ADDRESS)),
-            ''.join('-'*(COL_WIDTH_BYTE)),
-            ''.join('-'*(COL_WIDTH_BINARY)),
-        )
-        output.write(f'\n{header_text}\n{header_line_text}\n')
-        for lobj in line_obs:
-            line_str = f'{lobj.line_id.line_num}'.rjust(7)
-            address_value = lobj.address
-            address_str = address_format_str.format(address_value).center(COL_WIDTH_ADDRESS)
-            instruction_str = lobj.instruction.ljust(max_instruction_text_size)
-            if isinstance(lobj, LineWithBytes):
-                line_bytes = lobj.get_bytes()
-            else:
-                line_bytes = None
-            if line_bytes is not None:
-                bytes_list = list(line_bytes)
-                # print first line
-                output.write(
-                    f' {line_str} | {instruction_str} | {address_str} | 0x{bytes_list[0]:02x} | '
-                    f'{bytes_list[0]:08b} | {lobj.comment}\n'
-                )
-                for b in bytes_list[1:]:
-                    address_value += 1
-                    address_str = address_format_str.format(address_value).center(COL_WIDTH_ADDRESS)
-                    output.write(
-                        f' {blank_line_num} | {blank_instruction_text} | {address_str} | 0x{b:02x} | {b:08b} |\n'
-                    )
-            else:
-                output.write(
-                    f' {line_str} | {instruction_str} | {blank_address_text} | {blank_byte_text} | '
-                    f'{blank_binary_text} | {lobj.comment}\n'
-                )
-        return output.getvalue()