+++ bespokeasm/assembler/keywords.py
@@ -11,8 +11,15 @@ PREPROCESSOR_DIRECTIVES_SET = set([
     'include', 'require',
 ])
 
+EXPRESSION_FUNCTIONS_SET = set([
+    'LSB',
+] + [
+    f'BYTE{i}' for i in range(10)
+])
+
 ASSEMBLER_KEYWORD_SET = (
     COMPILER_DIRECTIVES_SET
     .union(BYTECODE_DIRECTIVES_SET)
     .union(PREPROCESSOR_DIRECTIVES_SET)
+    .union(EXPRESSION_FUNCTIONS_SET)
 )